Added methods needed for calculating a hash value for the grid used in querydata
NFmiQueryData objects use an internal cache for latlon values, which
speeds up many computations if the spatial reference is not
geographic. However, it may be quite expensive to initialize the
cache, especially for very large grids.

The Smartmet Server potentially loads several or even hundreds of
files from the same producer. The files for each producer are likely
to use the same grid. Hence the Smartmet Server caches the latlon
values for speed, and will set the values for newly loaded
NFmiQueryData objects from the cache so that they don't have to be
calculated again.

If the internal latlon cache is calculated from scratch for each new
querydata, say for highly requested data such as the global EC-model,
all would be on hold until the internal cache has been initialized by
the thread that got the first request. The queue may exhaust all
available threads and eventually even the queue of requests waiting
for an available threads. In practise this may lead to automatic
server restarts.

To also save memory, it is best to change the datatype of the internal
cache from a regular object to a shared one. For example typical radar
data takes 4 bytes per grid point in querydata format, but the
coordinates in NFmiPoint are of type double, and hence take 2*8 bytes,
four times as much as the data itself.

Unfortunately modifying the type of the cache changes the size of an
NFmiQueryData object, which means most modules using querydata
directly will have to be recompiled.

Being able to cache the coordinates greatly improved the speed of
Smartmet Server restarts, especially if the restart algorithm made
sure that there was a queue per producer making thus sure the
coordinates would have to be calculated only once.