Added an error check for an unknown string type
clang++ --analyze warning:

newbase/NFmiStringList.cpp:319:5: warning: Forming reference to null pointer
    file >> *item;
    ^~~~~~~~~~~~~

If the input stream is corrupted, the code could read a number not matching
any string type. Variable item would then remain unallocated, and dereferencing
the pointer would lead to a segmentation fault. The code now throws if the
string if is not recognized.